rename: rename folder component to components
