+++ src/components/PageDetail.tsx
